backend/requirements.txt
fastapi[standard]
pydantic
mysql-connector-python